+++ src/wxtilesdeckgl.ts
@@ -1,8 +1,12 @@
 import './wxtilesdeckgl.css';
 
-export { setupWxTilesLib, setWxTilesLogging, createWxTilesLayerProps, CreateProps, WxServerVarsStyleType } from './libtools/libTools';
-
+export { setupWxTilesLib, setWxTilesLogging, createWxTilesLayerProps } from './libtools/libTools';
+export type { CreateProps, WxServerVarsStyleType } from './libtools/libTools';
 export { WxTilesLayerManager, createDeckGlLayer } from './libtools/createDeckGlLayer';
-export { LibSetupObject, WXLOG } from './utils/wxtools';
-export { WxTilesLayer, WxTilesLayerProps } from './layers/WxTilesLayer';
+export type { LibSetupObject } from './utils/wxtools';
+export { WXLOG } from './utils/wxtools';
+export { WxTilesLayer } from './layers/WxTilesLayer';
+export type { WxTilesLayerProps } from './layers/WxTilesLayer';
 export { DebugTilesLayer } from './layers/DebugTilesLayer';
+export type { Legend } from './utils/RawCLUT';
+export { createLegend } from './utils/RawCLUT';